sempre usar long long a partir de agr

+++ graph.cpp
@@ -150,9 +150,9 @@ void toposort(){
 
 ///////KUHN'S ALGORITHM//////////////////////////////
 //bipartite unweighted graph maximum matching
-vector<int> g[200]; //grafo bipartido
+vector<ll> g[200]; //grafo bipartido
 vector<bool> vis(200);
-int match[200];
+ll match[200];
 //m vertices em um grupo, n vertices em outro
 void clrvis(){
     fill(vis.begin(),vis.end(),0);
@@ -160,10 +160,10 @@ void clrvis(){
 void clrmatch(){
     fill(match,match+n+m,-1);//só é necessario limpar o range utilizado nos matchs (m ou n)
 }
-bool kuhn(int u){//dfs based
+bool kuhn(ll u){//dfs based
     if(vis[u]) return 0;
     vis[u]=1;
-    for(int v : g[u]){
+    for(ll v : g[u]){
         if(match[v]==-1 or kuhn(match[v])){
             match[v]=u;
             return 1;
@@ -171,14 +171,15 @@ bool kuhn(int u){//dfs based
     }
     return 0;
 }
-int findMaxMatching(){
-    int ans=0, aux=0;
+ll findMaxMatching(){
+    ll ans=0, aux=0;
     clrmatch();//limpa array de matches
-    for(int i=n;i<n+m;i++){//dfs em todos os vertices do 2o grupo (n...n+m-1)
+    for(ll i=n;i<n+m;i++){//dfs em todos os vertices do 2o grupo (n...n+m-1)
         aux=0; 
         clrvis();//limpa os visitados
         kuhn(i);
-        for(int j=0;j<n;j++) aux += (match[j]!=-1);//conta os utilizados do 1o grupo (0..n-1)
+        for(ll j=0;j<n;j++) aux += (match[j]!=-1);//conta os utilizados do 1o grupo (0..n-1)
+		// {j, match[j]} eh um match
         ans=max(ans,aux);
     }
 return ans;
